Provide better Separation of Concerns
- extract classes to have a more readable Camel route
